DDP-5593 Modal activity block restructuring: - Move the component to its own directory - Move the styles and template to separate files - Add some css variables to customize the appearance of the component - Use studyGuid from input
