+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/modal-activity-block/modal-activity-block.component.scss
@@ -1,0 +1,132 @@
+.modal-activity-block {
+  line-height: 24px;
+
+  &__card {
+    padding: var(--modal-activity-block-card-padding, 18px 37px 18px 32px);
+  }
+
+  &__card-content {
+    display: flex;
+    flex-flow: row;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+  &__card-title, &__card-subtitle {
+    margin-bottom: 0;
+    color: var(--modal-activity-block-title-color, #1D1B25);
+    letter-spacing: 0;
+
+    font-size: var(--modal-activity-block-title-font-size, 20px);
+  }
+
+  &__card-subtitle {
+    color: var(--modal-activity-block-subtitle-color, #4A4951);
+    font-size: var(--modal-activity-block-subtitle-font-size, 16px);
+  }
+
+  &__card-buttons > button {
+    margin-left: 8px;
+    width: var(--modal-activity-block-button-size, 42px);
+    height: var(--modal-activity-block-button-size, 42px);
+  }
+
+  &__success-message {
+    font-size: var(--modal-activity-block-success-message-font-size, 14px);
+    color: var(--modal-activity-block-success-message-color, #03B486);
+
+    & mat-icon {
+      vertical-align: bottom;
+    }
+  }
+
+  &__save_activity {
+    font-size: 16px;
+    line-height: 20px;
+    width: 200px;
+    height: 48px;
+    padding: var(--modal-activity-block-save-activity-padding, 14px 0);
+    margin-top: 36px;
+  }
+
+  &__delete-dialog {
+    display: flex;
+    flex-flow: column;
+
+    &-title {
+      color: var(--modal-activity-block-title-color, #1D1B25);
+      font-size: 18px;
+      letter-spacing: 0;
+      text-align: center;
+      margin-bottom: 15px;
+      white-space: pre-line;
+    }
+
+    &-buttons {
+      display: inline-flex;
+      align-self: center;
+      gap: var(--modal-activity-block-delete-dialog-button-gap, 16px);
+
+      & > button {
+        height: var(--modal-activity-block-delete-dialog-button-height, 40px);
+        width: var(--modal-activity-block-delete-dialog-button-width, 140px);
+      }
+    }
+  }
+
+  &__edit-dialog {
+    &-title {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      padding-top: 30px;
+      padding-bottom: 20px;
+      margin: 0;
+      color: var(--modal-activity-block-title-color, #1D1B25);
+      font-size: var(--modal-activity-block-dialog-title-font-size, 24px);
+      letter-spacing: 0;
+      border-bottom: var(--modal-activity-block-dialog-title-border-bottom, 2px solid #E2E2E2);
+    }
+
+    &-content {
+      padding-bottom: 40px;
+      margin: 0;
+    }
+  }
+
+  &__close-button {
+    position: relative;
+    top: -5px;
+  }
+}
+
+::ng-deep {
+  & .modal-activity-block__edit-dialog {
+    & > mat-dialog-container {
+      padding: 0;
+
+      & > * {
+        box-sizing: border-box;
+        padding: 0 40px 0 93px;
+      }
+    }
+  }
+
+  & .modal-activity-block__delete-dialog {
+    & > mat-dialog-container {
+      padding: 27px 45px 20px 45px;
+    }
+
+    &:after { // arrow
+      content: " ";
+      position: relative;
+      height: 0;
+      right: 50%;
+      bottom: -100%;
+      border-top: 10px solid white;
+      border-right: 10px solid transparent;
+      border-left: 10px solid transparent;
+      border-bottom: none;
+    }
+  }
+}
