+++ ddp-workspace/projects/ddp-sdk/src/public-api.ts
@@ -123,7 +123,7 @@ export * from './lib/components/activityForm/questionPrompt.component';
 export * from './lib/components/activityForm/conditionalBlock.component';
 export * from './lib/components/activityForm/groupBlockList.component';
 export * from './lib/components/activityForm/groupBlock.component';
-export * from './lib/components/activityForm/modal-activity-block.component';
+export * from './lib/components/activityForm/modal-activity-block/modal-activity-block.component';
 export * from './lib/components/activityForm/activityContent.component';
 export * from './lib/components/activityForm/activityAgreementAnswer.component';
 export * from './lib/components/activityForm/activityCompositeAnswer.component';
